+++ packages/modules/app/src/app/index.ts
@@ -1,3 +1,3 @@
-export { App, IApp } from './App';
+export { App, IApp, type AppInitializeResult } from './App';
 
 export { default } from './App';