+++ packages/react/ag-grid/src/index.ts
@@ -1,4 +1,4 @@
-export { AgGridReact } from 'ag-grid-react';
+export { AgGridReact, AgGridReactProps } from 'ag-grid-react';
 
 export { enableAgGrid, type IAgGridProvider } from '@equinor/fusion-framework-module-ag-grid';